added authorization

+++ login/views.py
@@ -45,8 +45,8 @@ def register(request):
 		if user:
 			return JsonResponse({'error': 'A user with that username already exists.'}, status=409)
 		else:
-			user = User.objects.create(username=data['username'], password=hash(data['password']), role=data['role'])
-			return JsonResponse({"result":"success", "data":{'id': user.id,'name':user.username, 'role': user.role}}, status=201)
+			user = User.objects.create(username=data['username'], password=hash(data['password']), role="none")
+			return JsonResponse({"result":"success", "data":{'id': user.id, 'name':user.username}}, status=201)
 	else:
 		return JsonResponse({'error': 'This endpoint only accepts POST requests.'}, status=405)
 # Edit
@@ -56,13 +56,16 @@ def register(request):
 @csrf_exempt
 def edit(request, id):
 	if request.method == 'POST':
-		data = loads(request.body)
-		user = User.objects.filter(id=id)
-		if user:
-			user.update(username=data['username'], password=hash(data['password']), role=data['role'])
-			return JsonResponse({"result":"success","data":{'id': user[0].id, 'role': user[0].role, "hashcheckREMOVE":user[0].password}}, safe=False, status=200)
+		if request.session['user_role'] == 'admin' or request.session['user_id'] == id:
+			data = loads(request.body)
+			user = User.objects.filter(id=id)
+			if user:
+				user.update(username=data['username'], password=hash(data['password']), role=data['role'])
+				return JsonResponse({"result":"success","data":{'id': user[0].id, 'role': user[0].role, "hashcheckREMOVE":user[0].password}}, safe=False, status=200)
+			else:
+				return JsonResponse({'error': 'No user found with that id.'}, status=400)
 		else:
-			return JsonResponse({'error': 'No user found with that id.'}, status=400)
+			return JsonResponse({'error': 'You do not have permission to edit this user.'}, status=401)
 	else:
 		return JsonResponse({'error': 'This endpoint only accepts POST requests.'}, status=405)
 # Delete
@@ -72,12 +75,15 @@ def edit(request, id):
 @csrf_exempt
 def delete(request, id):
 	if request.method == 'POST':
-		user = User.objects.filter(id=id)
-		if user:
-			user.delete()
-			return JsonResponse({'result': 'success'}, status=200)
+		if request.session['user_role'] == 'admin' or request.session['user_id'] == id:
+			user = User.objects.filter(id=id)
+			if user:
+				user.delete()
+				return JsonResponse({'result': 'success'}, status=200)
+			else:
+				return JsonResponse({'error': 'No user found with that id.'}, status=400)
 		else:
-			return JsonResponse({'error': 'No user found with that id.'}, status=400)
+			return JsonResponse({'error': 'You do not have permission to delete this user.'}, status=401)
 	else:
 		return JsonResponse({'error': 'This endpoint only accepts POST requests.'}, status=405)
 # Hash
